Fixed Map edges
