+++ src/components/PresentationComponents/Map/NodeEdgesCurvedLinesMap.tsx
@@ -1,32 +1,27 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useMap } from 'react-leaflet';
 import L, { LatLngExpression, Marker } from 'leaflet';
 import 'leaflet.markercluster';
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css';
 import 'leaflet/dist/leaflet.css';
-import {
-  getMaxValueNode,
-  getMinValueNode,
-} from '@/utils/functions/getMinMaxValueNode';
 import { getNodeColorMapVoyagesStyle } from '@/utils/functions/getNodeColorStyle';
 import '@johnconnor_mulligan/leaflet.curve';
 import { RootState } from '@/redux/store';
 import { useSelector } from 'react-redux';
-import { maxRadiusInPixels, minRadiusInPixels } from '@/share/CONST_DATA';
 import {
   EdgesAggroutes,
-  EdgesAggroutedSourceTarget,
-  NodeAggroutes,
   CustomMarker,
   LatLng,
 } from '@/share/InterfaceTypesMap';
-import * as d3 from 'd3';
 import { getEdgesSize, getNodeSize } from '@/utils/functions/getNodeSize';
 import '@/style/map.scss';
-import { createSourceEdges } from './createSourceAndTargetDictionaries';
 import renderEdgesAnimatedLinesOnMap from './renderEdgesAnimatedLinesOnMap';
 import renderEdgesLinesOnMap from './renderEdgesLinesOnMap';
 import { createNodeDict } from '@/utils/functions/createNodeDict';
+import { handleHoverOriginMarkerCluster } from './handleHoverOriginMarkerCluster';
+import { handleHoverPostDisembarkationsMarkerCluster } from './handleHoverPostDisembarkationsMarkerCluster';
+import { createLogValueScale } from '@/utils/functions/createNodeLogValueScale';
+import { handleHoverCircleMarker } from './handleHoverCircleMarker';
 
 const NodeEdgesCurvedLinesMap = () => {
   const map = useMap();
@@ -35,23 +30,8 @@ const NodeEdgesCurvedLinesMap = () => {
     (state: RootState) => state.getNodeEdgesAggroutesMapData
   );
 
-  const nodesMap = new Map<string, NodeAggroutes>();
-
-  nodesData.forEach((node) => {
-    nodesMap.set(node.id, node);
-  });
-
-  const sourceEdgesMap = new Map<string, EdgesAggroutes[]>();
-  const targetEdgesMap = new Map<string, EdgesAggroutes[]>();
-  edgesData.forEach((edge) => {
-    if (!sourceEdgesMap.has(edge.source)) sourceEdgesMap.set(edge.source, []);
-    sourceEdgesMap.get(edge.source)?.push(edge);
-
-    if (!targetEdgesMap.has(edge.target)) targetEdgesMap.set(edge.target, []);
-    targetEdgesMap.get(edge.target)?.push(edge);
-  });
-  const hiddenEdgesLayer = L.layerGroup();
   const updateEdgesAndNodes = () => {
+    const hiddenEdgesLayer = L.layerGroup();
     map.eachLayer((layer) => {
       if (
         layer instanceof L.Curve ||
@@ -62,16 +42,7 @@ const NodeEdgesCurvedLinesMap = () => {
         map.removeLayer(layer);
       }
     });
-
-    const domainRanges = [
-      getMinValueNode(nodesData),
-      getMaxValueNode(nodesData),
-    ];
-
-    const nodeLogValueScale = d3
-      .scaleLog()
-      .domain(domainRanges)
-      .range([minRadiusInPixels, maxRadiusInPixels]);
+    const nodeLogValueScale = createLogValueScale(nodesData);
     const hiddenEdges = edgesData.filter(
       (edge) => edge.type === 'origination' || edge.type === 'disposition'
     );
@@ -145,49 +116,12 @@ const NodeEdgesCurvedLinesMap = () => {
         });
       },
     }).on('clustermouseover', async (event) => {
-      hiddenEdgesLayer.clearLayers();
-      const clusterLatLon = event.layer.getLatLng();
-      const clusterChildMarkers = event.layer.getAllChildMarkers();
-      const nodeIds = clusterChildMarkers.map(
-        (childMarker: any) => childMarker.nodeId
+      handleHoverOriginMarkerCluster(
+        event,
+        hiddenEdgesLayer,
+        hiddenEdges,
+        nodesData
       );
-
-      const targetNodeMap = new Map<string, [NodeAggroutes, EdgesAggroutes]>();
-
-      nodeIds.forEach((childNodeId: string) => {
-        hiddenEdges
-          .filter((edge) => edge.source === childNodeId)
-          .forEach((edge) => {
-            const nodes = nodesData.filter((node) => node.id === edge.target);
-            for (const node of nodes) {
-              targetNodeMap.set(node.id, [node, edge]);
-            }
-          });
-      });
-
-      for (const [, [node, edge]] of targetNodeMap) {
-        const { lat: clusterLat, lng: clusterLng } = clusterLatLon;
-        const { lat: nodeLat, lon: nodeLng } = node.data;
-        const size = getEdgesSize(edge);
-        const weightEddg = size !== null ? nodeLogValueScale(size) / 2 : 0;
-        const curveAnimated = renderEdgesAnimatedLinesOnMap(
-          [clusterLat, clusterLng],
-          [nodeLat!, nodeLng!],
-          weightEddg,
-          edge.controls
-        );
-        const curveLine = renderEdgesLinesOnMap(
-          [clusterLat, clusterLng],
-          [nodeLat!, nodeLng!],
-          weightEddg,
-          edge.controls,
-          edge.type
-        );
-        if (curveAnimated && curveLine) {
-          hiddenEdgesLayer.addLayer(curveLine);
-          hiddenEdgesLayer.addLayer(curveAnimated);
-        }
-      }
     });
 
     // Render postDisembarkationsMarkerCluster
@@ -222,97 +156,25 @@ const NodeEdgesCurvedLinesMap = () => {
         });
       },
     }).on('clustermouseover', (event) => {
-      hiddenEdgesLayer.clearLayers();
-      const clusterLatLon = event.layer.getLatLng();
-      const clusterChildMarkers = event.layer.getAllChildMarkers();
-      const nodeIdsClusters = clusterChildMarkers.map(
-        (childMarker: any) => childMarker.nodeId
+      handleHoverPostDisembarkationsMarkerCluster(
+        event,
+        hiddenEdgesLayer,
+        edgesData,
+        nodesData
       );
-
-      const sourceNodes: [NodeAggroutes, EdgesAggroutes][] = [];
-      const targetNodes: [NodeAggroutes, EdgesAggroutes][] = [];
-
-      for (const childNodeId of nodeIdsClusters) {
-        const targetEdges = targetEdgesMap.get(childNodeId);
-        if (targetEdges) {
-          for (const targetEdge of targetEdges) {
-            const sourceNode = nodesMap.get(targetEdge.source);
-            if (sourceNode) sourceNodes.push([sourceNode, targetEdge]);
-          }
-        }
-        const sourceEdges = sourceEdgesMap.get(childNodeId);
-        if (sourceEdges) {
-          for (const sourceEdge of sourceEdges) {
-            const targetNode = nodesMap.get(sourceEdge.target);
-            if (targetNode) targetNodes.push([targetNode, sourceEdge]);
-          }
-        }
-      }
-
-      const { lat: clusterLat, lng: clusterLng } = clusterLatLon;
-
-      // sourceNodes --> clusterNode to target
-      for (const [sourceNode, targetEdge] of sourceNodes) {
-        const { controls, type } = targetEdge;
-        const { lat: sourceLat, lon: sourceLng } = sourceNode.data!;
-
-        const size = getEdgesSize(targetEdge);
-        const weightEddg = size !== null ? nodeLogValueScale(size) / 2 : 0;
-
-        const curveAnimated = renderEdgesAnimatedLinesOnMap(
-          [sourceLat!, sourceLng!],
-          [clusterLat, clusterLng],
-          weightEddg,
-          controls
-        );
-        const curveLine = renderEdgesLinesOnMap(
-          [sourceLat!, sourceLng!],
-          [clusterLat, clusterLng],
-          weightEddg,
-          controls,
-          type
-        );
-        if (curveLine && curveAnimated) {
-          hiddenEdgesLayer.addLayer(curveLine);
-          hiddenEdgesLayer.addLayer(curveAnimated);
-        }
-      }
-
-      // targetNodes --> clusterNode to source
-      for (const [targetNode, sourceEdge] of targetNodes) {
-        const { controls, type } = sourceEdge;
-        const { lat: targetLat, lon: targetLng } = targetNode.data!;
-
-        const size = getEdgesSize(sourceEdge);
-        const weightEddg = size !== null ? nodeLogValueScale(size) / 2 : 0;
-        const curveAnimated = renderEdgesAnimatedLinesOnMap(
-          [targetLat!, targetLng!],
-          [clusterLat, clusterLng],
-          weightEddg,
-          controls
-        );
-        const curveLine = renderEdgesLinesOnMap(
-          [targetLat!, targetLng!],
-          [clusterLat, clusterLng],
-          weightEddg,
-          controls,
-          type
-        );
-
-        if (curveLine && curveAnimated) {
-          hiddenEdgesLayer.addLayer(curveLine);
-          hiddenEdgesLayer.addLayer(curveAnimated);
-        }
-      }
     });
 
-    const aggregatedEdges = new Map<string, EdgesAggroutedSourceTarget>();
     const originNodeMarkersMap = new Map<string, Marker>();
 
     nodesData.forEach((node) => {
       const { data, weights, id: nodeID } = node;
       const { lat, lon, name } = data;
-      const { origin, 'post-disembarkation': postDisembarkation } = weights;
+      const {
+        origin,
+        'post-disembarkation': postDisembarkation,
+        disembarkation,
+        embarkation,
+      } = weights;
       const size = getNodeSize(node);
       const nodeColor = getNodeColorMapVoyagesStyle(node);
       const logSize = nodeLogValueScale(size);
@@ -332,82 +194,30 @@ const NodeEdgesCurvedLinesMap = () => {
 
         const popupContent = `<p>${name}</p>`;
         circleMarker.bindPopup(popupContent);
-
-        circleMarker.on('mouseover', (event) => {
-          hiddenEdgesLayer.clearLayers();
-
-          const nodeHoverID = event.target.nodeId;
-
-          const hiddenEdgesData = edgesData.filter(
-            (edge) =>
-              (edge.target === nodeHoverID || edge.source === nodeHoverID) &&
-              (edge.type === 'origination' || edge.type === 'disposition')
+        circleMarker.on('mousemove', (event) => {
+          handleHoverCircleMarker(
+            event,
+            hiddenEdgesLayer,
+            edgesData,
+            nodesData,
+            originNodeMarkersMap,
+            originMarkerCluster
           );
-
-          const sourceEdges = createSourceEdges(nodeHoverID, hiddenEdgesData);
-
-          const targetNode = nodesData.find((node) => node.id === nodeHoverID)!;
-
-          const { lat: targetLat, lon: targetLng } = targetNode?.data!;
-          sourceEdges.forEach((sourceEdge) => {
-            const sourceNodeId = sourceEdge.source;
-            const originNode = originNodeMarkersMap.get(sourceNodeId);
-            const visibleParent = originMarkerCluster.getVisibleParent(
-              originNode!
-            );
-
-            const { lat: sourceLat, lng: sourceLng } =
-              visibleParent.getLatLng();
-
-            const parentKeyLeaflet = [sourceLat, sourceLng].join();
-
-            if (!aggregatedEdges.has(parentKeyLeaflet)) {
-              const newAggregatedEdge: EdgesAggroutedSourceTarget = {
-                ...sourceEdge,
-                sourceLatlng: [sourceLat, sourceLng],
-                targetLatlng: [targetLat!, targetLng!],
-                controls: sourceEdge.controls,
-                weight: sourceEdge.weight,
-              };
-              aggregatedEdges.set(parentKeyLeaflet, newAggregatedEdge);
-            }
-          });
-
-          for (const [, edgeData] of aggregatedEdges) {
-            const { sourceLatlng, targetLatlng, controls, type } = edgeData;
-
-            const size = getEdgesSize(edgeData);
-            const weightEddg =
-              size !== null ? nodeLogValueScale(size) / 1.5 : 0;
-
-            const curveAnimated = renderEdgesAnimatedLinesOnMap(
-              sourceLatlng,
-              targetLatlng,
-              weightEddg,
-              controls
-            );
-            const curveLine = renderEdgesLinesOnMap(
-              targetLatlng,
-              sourceLatlng,
-              weightEddg,
-              controls,
-              type
-            );
-            if (curveAnimated && curveLine) {
-              hiddenEdgesLayer.addLayer(curveLine);
-              hiddenEdgesLayer.addLayer(curveAnimated);
-            }
-          }
         });
 
-        if (origin && origin > 0) {
+        if (disembarkation !== 0 || embarkation !== 0) {
+          circleMarker.addTo(map).bringToFront();
+        } else if (origin && origin > 0) {
           originNodeMarkersMap.set(nodeID, originMarker);
           originMarkerCluster.addLayer(circleMarker);
           originMarkerCluster.addLayer(originMarker);
-        } else if (postDisembarkation && postDisembarkation > 0) {
+        } else if (
+          Number(postDisembarkation) &&
+          Number(postDisembarkation) > 0 &&
+          disembarkation === 0 &&
+          embarkation === 0
+        ) {
           postDisembarkationsMarkerCluster.addLayer(circleMarker);
-        } else {
-          circleMarker.addTo(map).bringToFront();
         }
       }
     });
